Clean code into class
